the last image commit nchlah

+++ routes/profile.js
@@ -4,18 +4,25 @@ const router = express.Router();
 const Profile = require("../models/Profile");
 //Require the User Schema
 const User=require("../models/User")
+//require image scheme
 const imgModel=require("../models/Image")
 
-const isAuth=require('../middlewares/isAuth')
-
+const {isAuth,isAdmin}=require('../middlewares/isAuth')
 
 
 
 // get one profile 
 router.get("/one/:_id", isAuth,async (req, res) => {
   const { _id } = req.params;
+  const userId =req.user._id
+  
   try {
-    const profile = await Profile.findById({_id});
+    const profileuUser = await Profile.findOne({user: userId} );
+    if (profileuUser._id!=_id){
+
+       return res.status(400).json({ msg: 'u dont have acces to view this profile' });
+
+    }    const profile = await Profile.findById({_id});
     res.json({ msg: 'profile ', profile });
 
   } catch (error) {
@@ -36,9 +43,17 @@ router.get("/", isAuth,async (req, res) => {
   }
 });
 // edit a profile 
-router.put("/edit/:_id", isAuth ,async (req, res) => {
+router.put("/edit/:_id", isAuth ,isAdmin,async (req, res) => {
   const { _id } = req.params;
+  const userId =req.user._id
   try {
+    const profileuUser = await Profile.findOne({user: userId} );
+    if (profileuUser._id!=_id){
+
+       return res.status(400).json({ msg: 'u dont have acces to edit this profile' });
+
+    }
+
     const profile = await Profile.findOneAndUpdate({ _id }, { $set: req.body }, {new:true});
     res.json({ msg: `profile edited `, profile });
   } catch (error) {
@@ -48,9 +63,16 @@ router.put("/edit/:_id", isAuth ,async (req, res) => {
 
 
   // delete a profile 
-  router.delete("/delete/:id", isAuth,async (req, res) => {
+  router.delete("/delete/:id", isAuth,isAdmin,async (req, res) => {
     const { id } = req.params;
+    const userId =req.user._id
     try {
+      const profileuUser = await Profile.findOne({user: userId} );
+      if (profileuUser._id!=id){
+
+         return res.status(400).json({ msg: 'u dont have acces to delete this profile' });
+ 
+      }
       const profile = await Profile.findOneAndDelete({  id });
       res.json({ msg: "profile deleted", profile });
     } catch (error) {
@@ -64,9 +86,18 @@ router.put("/edit/:_id", isAuth ,async (req, res) => {
 router.post("/addprofile/:index", isAuth, async (req, res) => {
   const {index}=req.params
   const {userName}= req.body
+  const _id=req.user._id
     try {
-      
-      const user = await User.findById(req.user._id).select("-password");
+      let profil = await Profile.findOne( {user :_id } );
+      let profilName = await Profile.findOne(  {userName} );
+      if (profil) {
+        return res.status(400).json({ msg: 'you have already a profil' });
+      }
+      if (profilName) {
+        return res.status(400).json({ msg: 'userName exisits' });
+      }
+      const user = await User.findById(_id).select("-password");
+      console.log(user)
       const image = await imgModel.findById(index);
    
         const newProfile = {
